requirements.txt
argon2-cffi @ file:///D:/bld/argon2-cffi_1605217174718/work
async-generator==1.10
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1605083924122/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache==1.6.1
bezier==2020.5.19
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1612213472466/work
certifi==2020.12.5
cffi @ file:///D:/bld/cffi_1613414034267/work
cloudpickle @ file:///home/conda/feedstock_root/build_artifacts/cloudpickle_1598400192773/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
cycler==0.10.0
cytoolz==0.11.0
dask @ file:///home/conda/feedstock_root/build_artifacts/dask-core_1614995065708/work
decorator==4.4.2
defusedxml @ file:///home/conda/feedstock_root/build_artifacts/defusedxml_1615232257335/work
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
h5py==2.8.0
imagecodecs @ file:///D:/bld/imagecodecs_1614194834651/work
imageio @ file:///tmp/build/80754af9/imageio_1594161405741/work
imgaug==0.4.0
importlib-metadata @ file:///D:/bld/importlib-metadata_1615778985732/work
ipykernel @ file:///D:/bld/ipykernel_1613853146220/work/dist/ipykernel-5.5.0-py3-none-any.whl
ipython @ file:///D:/bld/ipython_1614393754223/work
ipython-genutils==0.2.0
ipywidgets @ file:///home/conda/feedstock_root/build_artifacts/ipywidgets_1609995587151/work
jedi @ file:///D:/bld/jedi_1610146981085/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1612119311452/work
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema_1614815863336/work
jupyter==1.0.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1615693636836/work
jupyter-console @ file:///home/conda/feedstock_root/build_artifacts/jupyter_console_1615926265198/work
jupyter-core @ file:///D:/bld/jupyter_core_1612125436808/work
jupyterlab-pygments @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_pygments_1601375948261/work
jupyterlab-widgets @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_widgets_1609173350931/work
kiwisolver @ file:///D:/bld/kiwisolver_1610099949969/work
MarkupSafe @ file:///D:/bld/markupsafe_1610127696479/work
matplotlib @ file:///D:/bld/matplotlib-suite_1604871176074/work
mistune @ file:///D:/bld/mistune_1610113018274/work
mkl-fft==1.3.0
mkl-random==1.2.0
mkl-service==2.3.0
natsort==7.1.0
nbclient @ file:///home/conda/feedstock_root/build_artifacts/nbclient_1614336084111/work
nbconvert @ file:///D:/bld/nbconvert_1605401997174/work
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1611005694671/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1605195931949/work
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1598210780226/work
notebook @ file:///D:/bld/notebook_1610575470050/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1603468620949/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-python==4.5.1.48
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1612459636436/work
pandocfilters==1.4.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1607618318316/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow @ file:///D:/bld/pillow_1615244129446/work
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1605543085815/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1615486796966/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pygame==2.0.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1615243893546/work
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent @ file:///D:/bld/pyrsistent_1610146997928/work
python-dateutil==2.8.1
PyWavelets @ file:///D:/bld/pywavelets_1607290963729/work
pywin32==300
pywinpty @ file:///D:/bld/pywinpty_1602377770565/work
PyYAML==5.4.1
pyzmq @ file:///D:/bld/pyzmq_1614611876813/work
qtconsole @ file:///home/conda/feedstock_root/build_artifacts/qtconsole_1615946127199/work
QtPy==1.9.0
scikit-image==0.17.2
scipy @ file:///C:/bld/scipy_1615956095913/work
Send2Trash==1.5.0
Shapely==1.7.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
terminado @ file:///D:/bld/terminado_1616072422022/work
testpath==0.4.4
tifffile @ file:///home/conda/feedstock_root/build_artifacts/tifffile_1616017175199/work
toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1600973991856/work
tornado @ file:///D:/bld/tornado_1610094878337/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1602771532708/work
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1602702424206/work
UnrealCV==0.4.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
widgetsnbextension @ file:///D:/bld/widgetsnbextension_1605475785107/work
wincertstore==0.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1614945704755/work
